Viikkotehtävä4/requirements.txt
requests
mysql-connector-python
pandas
python-dotenv
streamlit
